+++ flames/ase_utils.py
@@ -223,7 +223,9 @@ def crystalOptimization(
     Space Group Number: {}
     Space Group Symbol: {}
     Lattice type: {}
-    """.format(symm.number, symm.international, atoms.cell.get_bravais_lattice().longname),
+    """.format(
+                    symm.number, symm.international, atoms.cell.get_bravais_lattice().longname
+                ),
                 file=out_file,
                 flush=True,
             )